add GetRunRef from the RunsClient
